+++ doc/md/Changelog.md
@@ -7,7 +7,7 @@ Changes
 - The jsoncons `boost` extension has been removed. That extension contained a sample `json_type_traits` specialization for `boost::gregorian::date`, which may still be found in the "Type Extensibility" tutorial.  
 - The member `json_type_traits` member function `assign` has been removed and replaced by `to_json`. if you have implemented your own type specializations, you will also have to change your `assign` function to `to_json`.
 - `json_type_traits` specializations no longer require the `is_assignable` data member
-- The names `json_deserializer`,`ojson_deserializer`,`wjson_deserializer`,`wojson_deserializer` have been deprecated (they still work) and renamed to `json_encoder`,`ojson_encoder`,`wjson_encoder`,`wojson_encoder`.  
+- The names `json_deserializer`,`ojson_deserializer`,`wjson_deserializer`,`wojson_deserializer` have been deprecated (they still work) and replaced by `template <class Json> json_encoder`.  
 - The name `output_format` has been deprecated (it still works) and renamed to `serialization_options`.  
 
 New features